New tick encoding using delta ticks, and new variable length encoding
This variable length encoding is inspired by what WebSocket's do.

Tick deltas 253 and lower use a single byte, if the length is between 254 and 0xFF then the first byte is written out as 254 and the length data is written out as a uint16, and if the length is greater than 0xFF the first byte is written as 255 and the full data then written out as a uint32 after

Doing extensive benchmarks this results in similar on disk size savings as the midi vlv encoding under most cases, but is up to 20% faster in the worst case for parsing performance.

The parsing code for this system is much simpler not requiring any bitmasking or loops just a single switch statement.

Another advantage is when looking at the byte stream under a hex editor it's much more obvious what is happening since 0xFF and 0xFE stand out a lot and make this scheme more human readable for people writing parsers.